NEW : Add dup_process_child() to handle I/O redirections in piped processes
This function sets up the correct file descriptors for a child process
in a pipeline, duplicating either standard input/output from existing
pipe FDs or from input/output redirection FDs if specified.

It handles the following cases:
- Redirects STDIN from the previous pipe or an input file descriptor
- Redirects STDOUT to the next pipe or an output file descriptor
- Closes unnecessary FDs to avoid leaks

This is part of the pipe and redirection management logic in the minishell.